==> Lessons/06_Lab_GLMs.Rmd <==
---
title: "6: Lab - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2023"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   
library(here)
here()

EPAair <- read.csv(here("Data/Processed_KEY/EPAair_O3_PM25_NC1819_Processed.csv"), stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv(here("Data/Processed_KEY/NEON_NIWO_Litter_mass_trap_Processed.csv"), stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r visualization}
#Exercise 1:
ozone.dens.plot <- ggplot(EPAair, aes(x = Ozone, fill = as.factor(Year))) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c('darkgreen', 'darkblue')) +
  labs(x = 'Ozone AQI Values', y = 'Density', fill = '')
  
print(ozone.dens.plot)

#2.
ozone.box.plot <- ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_boxplot() +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 4, label = c("b", "a")) +
  labs(y = 'Ozone AQI Values')
  
print(ozone.box.plot)

#3
ozone.violin.plot <- ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_violin(draw_quantiles = 0.5) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 4, label = c("b", "a")) +
  labs(y = 'Ozone AQI Values') +
  ylim(0,150)

print(ozone.violin.plot)
```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r linear regression}

#Exercise 2: Run a linear regression PM2.5 (y) by Ozone (x). Find the p-value and R-squared value. 

PM.ozone.regression <- lm(EPAair$PM2.5 ~ EPAair$Ozone)
summary(PM.ozone.regression)

#P - value = 2.2e-16, R-squared value = 0.1314

#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot
PM.by.ozone <- 
  ggplot(EPAair, aes(x = Ozone, y = PM2.5)) +
  geom_point() +
  geom_smooth(method = "lm", col = 'blue', se = FALSE)
#se = false gets rid of the gray halo of standard error that surrounds your line of best fit

print(PM.by.ozone)
  

```


## AIC to select variables

What other variables can we add to improve model?

```{r AIC}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.
library(corrplot)
EPAair.subset <- 
  EPAair %>%
  select(mean_Lat:Ozone) %>%
  drop_na()
#you cannot have NAs when trying to compute correlation

EPAairCorr <- cor(EPAair.subset)
corrplot.mixed(EPAairCorr, upper = "ellipse")
#EPAaircorr creates a matrix (table)

#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 4?

Ozone.ALL <- lm(data=EPAair.subset, PM2.5 ~ Ozone + Year + Month + mean_Lat + mean_Lng)
summary(Ozone.ALL)
step(Ozone.ALL)

#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 2. 
Ozone.best <- lm(data=EPAair.subset, PM2.5 ~ Ozone + Month + mean_Lat + mean_Lng)
summary(Ozone.best)

```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter.Totals %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot


```
